requirements.txt
streamlit
pymupdf
python-docx
deep-translator
fitz